+++ src/components/Note.jsx
@@ -1,9 +1,13 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import { useState, useEffect } from "react";
-import { updateDoc, getFirestore, doc, getDoc, documentId, deleteDoc, setDoc} from "firebase/firestore";
+import { updateDoc, getFirestore, doc, deleteDoc} from "firebase/firestore";
+import Swal from "sweetalert2";
+import withReactContent from "sweetalert2-react-content";
 
 function Note(props){
+
+    const MySwal = withReactContent(Swal)
     const navigate = useNavigate();
 
     const [savedNotes, setSavedNotes] = useState(props.notes);
@@ -19,6 +23,7 @@ function Note(props){
     
     let note = props.notes.find(note => note.title === title);
     let noteIdx = props.notes.indexOf(note);
+
 
     const db = getFirestore();
 
@@ -66,18 +71,33 @@ function Note(props){
             }
         }
     }
-    const dispose = () => {
-        if(origin === "/fav"){
-            let docRef =  doc(db, 'favs', props.note.title);
-            deleteDoc(docRef);
-            navigate('/', {replace:true})
-        }else{
-            let docRef =  doc(db, 'notes', props.note.title);
-            deleteDoc(docRef);
-            navigate('/', {replace:true})
-        }
-    }
 
+    const dispose = () => {
+        MySwal.fire({
+            customClass: {
+                confirmButton: "confirm-btn",
+                popup : "swal-cont"
+            },
+            title : '¿🗑️?',
+            showCancelButton: true,
+            confirmButtonText : <svg xmlns="http://www.w3.org/2000/svg" width="1rem" height="1rem" fill="#fff" viewBox="0 0 24 24"><path d="M9 22l-10-10.598 2.798-2.859 7.149 7.473 13.144-14.016 2.909 2.806z"/></svg> ,
+            cancelButtonText :<svg xmlns="http://www.w3.org/2000/svg" width="1rem" height="1rem" fill="#fff"viewBox="0 0 24 24"><path d="M23 20.168l-8.185-8.187 8.185-8.174-2.832-2.807-8.182 8.179-8.176-8.179-2.81 2.81 8.186 8.196-8.186 8.184 2.81 2.81 8.203-8.192 8.18 8.192z"/></svg> ,
+            denyButtonText: `Don't save`
+        })
+        .then( (result) =>{
+            if(result.isConfirmed){
+                if(origin === "/fav"){
+                    let docRef =  doc(db, 'favs', note.title);
+                    deleteDoc(docRef);
+                    navigate('/', {replace:true})
+                }else{
+                    let docRef =  doc(db, 'notes', note.title);
+                    deleteDoc(docRef);
+                    navigate('/', {replace:true})
+                }
+            }
+    })}
+            
     while(note !== null & note !== undefined){
         return(
             <>
